+++ Util/pzchunk.h
@@ -51,7 +51,7 @@ class TPZChunkVector
 	 * @brief Constructor with numberofchuncks chuncks.
 	 * @param numberofchunks Indicates how large the initial chunk vector will be.
 	 */
-	TPZChunkVector(int numberofchunks = DEFAULTNUMBEROFCHUNKS);
+	TPZChunkVector(int64_t numberofchunks = DEFAULTNUMBEROFCHUNKS);
 	/** @brief Destructor. */
 	virtual ~TPZChunkVector();
 	
@@ -59,14 +59,14 @@ class TPZChunkVector
 	 * @brief Access method to query the number of elements of the vector.
 	 * @return Returns the number of elements of the vector.
 	 */
-	inline int NElements() const {return fNElements;}
+	inline int64_t NElements() const {return fNElements;}
 	
 	/**
 	 * @brief Increase the size of the chunk vector.
 	 * @param newsize New size of the vector. Does not indicate how
 	 * much memory will be allocated!
 	 */
-	void Resize(const int newsize);
+	void Resize(const int64_t newsize);
 	
 	/** @brief Returns a reference to the ith element of the vector. */
 	/**
@@ -75,14 +75,14 @@ class TPZChunkVector
 	 * code exits if the index is out of bounds.  Note that NODEBUG
 	 * may modify the result of the code.
 	 */
-	T &operator[](const int nelem) const;
+	T &operator[](const int64_t nelem) const;
 	
 	/** @brief Finds the index of an object by its pointer */
-	int FindObject(T *object);
+	int64_t FindObject(T *object);
 	
 protected:
 	/** @brief Number of elements of the chunk vector. */
-	int fNElements;
+	int64_t fNElements;
 	
 	/** @brief Vector which points to each chunk of objects. */
 	TPZManVector<T*> fVec;
@@ -91,12 +91,12 @@ protected:
 //--| IMPLEMENTATION |----------------------------------------------------------
 
 template< class T, int EXP >
-TPZChunkVector<T,EXP>::TPZChunkVector(int numberofchunks) :
+TPZChunkVector<T,EXP>::TPZChunkVector(int64_t numberofchunks) :
 fVec(numberofchunks)
 {
 	fNElements = 0;
 	
-	for(int i=0; i<numberofchunks; i++)
+	for(int64_t i=0; i<numberofchunks; i++)
 		fVec[i] = 0;
 	
 	fVec.Resize(0);
@@ -105,16 +105,16 @@ fVec(numberofchunks)
 template< class T,int EXP >
 TPZChunkVector<T,EXP>::~TPZChunkVector()
 {
-	int nchunks=fVec.NElements();
+	int64_t nchunks=fVec.NElements();
 	
-	for(int i=0;i<nchunks;i++)
+	for(int64_t i=0;i<nchunks;i++)
 		if(fVec[i]) delete[] fVec[i];
 }
 
 // Increase the size of the chunk vector
 
 template< class T ,int EXP>
-void TPZChunkVector<T,EXP>::Resize(const int newsize) {
+void TPZChunkVector<T,EXP>::Resize(const int64_t newsize) {
 #ifndef NODEBUG
 	if(newsize<0) {
 		PZError << "TPZChunkVector::Resize. Bad parameter newsize." << std::endl;
@@ -124,10 +124,10 @@ void TPZChunkVector<T,EXP>::Resize(const int newsize) {
 #endif
 	fNElements = newsize;
 	
-	int i;
-	int nchunks = fVec.NElements();
+	int64_t i;
+	int64_t nchunks = fVec.NElements();
 	
-	int chunksneeded = ((newsize-1) >> EXP)+1;
+	int64_t chunksneeded = ((newsize-1) >> EXP)+1;
 	if (chunksneeded == nchunks) return;
 	
 	T *NullPointer = 0;
@@ -135,7 +135,7 @@ void TPZChunkVector<T,EXP>::Resize(const int newsize) {
 	if(chunksneeded > nchunks)
 		fVec.Resize(chunksneeded,NullPointer);
 	
-	const int sizechunk = 1 << EXP;
+	const int64_t sizechunk = 1 << EXP;
 	for(i = 0; i<chunksneeded; i++ )
 		if(!fVec[i])
 			fVec[i] = new T[sizechunk];
@@ -149,12 +149,12 @@ void TPZChunkVector<T,EXP>::Resize(const int newsize) {
 }
 
 template<class T, int EXP>
-int TPZChunkVector<T,EXP>::FindObject(T *obj) {
-	int nch = fVec.NElements();
-	int ich;
-	int index = 0;
+int64_t TPZChunkVector<T,EXP>::FindObject(T *obj) {
+	int64_t nch = fVec.NElements();
+	int64_t ich;
+	int64_t index = 0;
 	// number of elements in a chunk
-	int nelch = 1<<EXP;
+	int64_t nelch = 1<<EXP;
 	for(ich=0; ich<nch; ich++) {
 		if(fVec[ich] == 0) continue;
 		if(obj >= fVec[ich] && obj < fVec[ich]+nelch) {
@@ -169,7 +169,7 @@ int TPZChunkVector<T,EXP>::FindObject(T *obj) {
 
 // Return a reference to the ith element of the vector
 template< class T, int EXP >
-inline T &TPZChunkVector<T,EXP>::operator[](const int nelem) const {
+inline T &TPZChunkVector<T,EXP>::operator[](const int64_t nelem) const {
 #ifndef NODEBUG
 	if(nelem<0 || nelem >= NElements()) {
 		PZError << "TPZChunkVector::operator[]. "
@@ -182,7 +182,7 @@ inline T &TPZChunkVector<T,EXP>::operator[](const int nelem) const {
 	}
 #endif
 	
-	const int mask = (1 << EXP)-1;
+	const int64_t mask = (1 << EXP)-1;
 	
 	// nelem & mask deve ser nelem ??
 	return (fVec[nelem >> EXP])[nelem & mask];
@@ -193,13 +193,13 @@ TPZChunkVector<T,EXP>::TPZChunkVector(const TPZChunkVector<T,EXP> &TCh) :
 fVec(TCh.fVec.NElements()) {
 	
 	fNElements = TCh.NElements();
-	int nchunks = TCh.fVec.NElements();
+	int64_t nchunks = TCh.fVec.NElements();
 	
-	for(int i=0; i<nchunks; i++) {
+	for(int64_t i=0; i<nchunks; i++) {
 		if(!TCh.fVec[i]) {
 			fVec[i] = 0;
 		} else {
-			int j,k=1<<EXP;
+			int64_t j,k=1<<EXP;
 			T* ptr = new  T[1 << EXP];
 			fVec[i] = ptr;
 			T* ptrcp = TCh.fVec[i];
@@ -211,9 +211,9 @@ fVec(TCh.fVec.NElements()) {
 template < class T, int EXP >
 TPZChunkVector<T,EXP> & TPZChunkVector<T,EXP>::operator=(const TPZChunkVector<T,EXP> &TCh){
 	if(this == &TCh) return *this;
-	int prvsz = fVec.NElements();
-	int sz = TCh.fVec.NElements();
-	int i;
+	int64_t prvsz = fVec.NElements();
+	int64_t sz = TCh.fVec.NElements();
+	int64_t i;
 	for(i=sz;i<prvsz;i++)
 		if(fVec[i])
 			delete[] fVec[i];
@@ -222,13 +222,13 @@ TPZChunkVector<T,EXP> & TPZChunkVector<T,EXP>::operator=(const TPZChunkVector<T,
 	fVec.Shrink();
 	fNElements = TCh.NElements();
 	
-	int nchunks = sz;
+	int64_t nchunks = sz;
 	
 	for(i=0; i<nchunks; i++) {
 		if(!TCh.fVec[i]) {
 			fVec[i] = 0;
 		} else {
-			int j,k=1<<EXP;
+			int64_t j,k=1<<EXP;
 			fVec[i] = new T[1 << EXP];
 			
 			for(j=0;j<k;j++)
